Splitting out all but status from iree/base/api.h.
